requirements.txt
# AWS Lambda DICOM to JPG Converter Requirements

# Core dependencies
boto3>=1.34.0
botocore>=1.34.0

# DICOM handling
pydicom>=2.4.4

# Image processing
Pillow>=10.4.0
numpy>=1.24.0

# HTTP requests
requests>=2.31.0

# Additional dependencies that might be needed
urllib3>=1.26.18
python-dateutil>=2.8.2
jmespath>=1.0.1
s3transfer>=0.10.0

# Type hints support
typing-extensions>=4.9.0 